core: move EigenToPy to a dedicated file
