+++ CMakeLists.txt
@@ -107,6 +107,7 @@ SET(${PROJECT_NAME}_HEADERS
   include/eigenpy/details.hpp
   include/eigenpy/fwd.hpp
   include/eigenpy/eigen-allocator.hpp
+  include/eigenpy/eigen-to-python.hpp
   include/eigenpy/map.hpp
   include/eigenpy/geometry.hpp
   include/eigenpy/geometry-conversion.hpp
